requirements.txt
Flask
flask-autodoc
tornado
flask-cors
Flask-Script
Flask-Security
Flask-Reggie
python-json-logger
flask-paginate
requests
psycopg2
pyyaml